inject backstoptools
