+++ backstop_data/engine_scripts/goToNextPage.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = (chromy, scenario, vp) => {
+    chromy.goto('https://github.com/garris/BackstopJS/issues');
+};
